Added Windowing System
This adds the basic windowing system for Nunya. The windowing system
provides wrappers for graphics drawing, and enforces only drawing within
a window bounds. This is accomplished by adding a new feature to the
graphics system which is an optional restrcited range. Windows use this
to enforce clipping in a simple manner, and then clear it when finished.
It is also used to ensure we are not attempting to draw offscreen and
therefore not accessing other parts of memory.